Consolidate entry points (#1761)
