+++ package.json
@@ -13,7 +13,7 @@
   "repository": "github:reduxjs/react-redux",
   "bugs": "https://github.com/reduxjs/react-redux/issues",
   "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
+  "types": "./es/index.d.ts",
   "unpkg": "dist/react-redux.js",
   "module": "es/index.js",
   "files": [
