+++ src/alternate-renderers.ts
@@ -1,0 +1,9 @@
+export * from './exports'
+
+import { getBatch } from './utils/batch'
+
+// For other renderers besides ReactDOM and React Native,
+// use the default noop batch function
+const batch = getBatch()
+
+export { batch }
