+++ src/index.ts
@@ -1,66 +1,10 @@
-import Provider from './components/Provider'
-import type { ProviderProps } from './components/Provider'
-import connectAdvanced from './components/connectAdvanced'
-import type {
-  ConnectAdvancedOptions,
-  ConnectProps,
-} from './components/connectAdvanced'
-import type {
-  SelectorFactory,
-  Selector,
-  MapStateToProps,
-  MapStateToPropsFactory,
-  MapStateToPropsParam,
-  MapDispatchToPropsFunction,
-  MapDispatchToProps,
-  MapDispatchToPropsFactory,
-  MapDispatchToPropsParam,
-  MapDispatchToPropsNonObject,
-  MergeProps,
-} from './connect/selectorFactory'
-import { ReactReduxContext } from './components/Context'
-import type { ReactReduxContextValue } from './components/Context'
-import connect from './connect/connect'
+export * from './exports'
 
-import { useDispatch, createDispatchHook } from './hooks/useDispatch'
-import { useSelector, createSelectorHook } from './hooks/useSelector'
-import { useStore, createStoreHook } from './hooks/useStore'
-
-import { setBatch } from './utils/batch'
 import { unstable_batchedUpdates as batch } from './utils/reactBatchedUpdates'
-import shallowEqual from './utils/shallowEqual'
+import { setBatch } from './utils/batch'
 
+// Enable batched updates in our subscriptions for use
+// with standard React renderers (ReactDOM, React Native)
 setBatch(batch)
 
-export * from './types'
-export type {
-  ProviderProps,
-  SelectorFactory,
-  Selector,
-  MapStateToProps,
-  MapStateToPropsFactory,
-  MapStateToPropsParam,
-  ConnectProps,
-  ConnectAdvancedOptions,
-  MapDispatchToPropsFunction,
-  MapDispatchToProps,
-  MapDispatchToPropsFactory,
-  MapDispatchToPropsParam,
-  MapDispatchToPropsNonObject,
-  MergeProps,
-  ReactReduxContextValue,
-}
-export {
-  Provider,
-  connectAdvanced,
-  ReactReduxContext,
-  connect,
-  batch,
-  useDispatch,
-  createDispatchHook,
-  useSelector,
-  createSelectorHook,
-  useStore,
-  createStoreHook,
-  shallowEqual,
-}
+export { batch }
